#23 Set MongoDB database for the project and build Patient Schema
