+++ server/package.json
@@ -13,7 +13,9 @@
   "license": "ISC",
   "dependencies": {
     "cors": "^2.8.5",
+    "dotenv": "^16.4.5",
     "express": "^4.21.0",
+    "mongoose": "^8.7.0",
     "morgan": "^1.10.0",
     "newman": "^6.2.1"
   },
