--- server/src/types/types.ts
@@ -1,8 +1,0 @@
-export interface PatientData {
-    data_qual_str : string[],
-    data_qual_time : number[],
-    measurement_data : number[],
-    start_date_time : string,
-    time_vector : number[],
-    units : string[]
-}
